Add section.
