+++ crates/console/src/components/mod.rs
@@ -2,10 +2,12 @@ mod entry;
 mod layout;
 mod list;
 mod map;
+mod section;
 mod styled_text;
 
 pub use entry::*;
 pub use layout::*;
 pub use list::*;
 pub use map::*;
+pub use section::*;
 pub use styled_text::*;
