+++ crates/console/src/ui.rs
@@ -4,18 +4,37 @@ use iocraft::prelude::*;
 use miette::IntoDiagnostic;
 
 pub use crate::components::*;
+pub use crate::theme::*;
 
 impl<R: Reporter> Console<R> {
     pub fn render<T: Component>(&self, element: Element<'_, T>) -> miette::Result<()> {
+        let theme = ConsoleTheme::default();
+
         self.out.flush()?;
-        element.into_any().print();
+
+        element! {
+            ContextProvider(value: Context::owned(theme)) {
+                #(element)
+            }
+        }
+        .print();
 
         Ok(())
     }
 
     pub async fn render_loop<T: Component>(&self, element: Element<'_, T>) -> miette::Result<()> {
+        let theme = ConsoleTheme::default();
+
         self.out.flush()?;
-        element.into_any().render_loop().await.into_diagnostic()?;
+
+        element! {
+            ContextProvider(value: Context::owned(theme)) {
+                #(element)
+            }
+        }
+        .render_loop()
+        .await
+        .into_diagnostic()?;
 
         Ok(())
     }
